+++ scripts/speculative_inference.py
@@ -172,7 +172,7 @@ def infer(ids):
 
     result, n_steps = speculative_generate(
         model,
-        ids,
+        ids.int(),
         speculator,
         new_tokens=100,
         max_seq_len=max_seq_len,
